scripts/sagemaker/inference_model/requirements.txt
# VapeWatch - Inference Requirements (SageMaker Compatible)

# Core
ultralytics==8.1.0
numpy>=1.23.0,<1.27
opencv-python-headless>=4.8.0
Pillow>=9.5.0,<11.0.0
matplotlib>=3.7.0

# AWS and SageMaker
boto3
sagemaker
botocore

# Utility
pyyaml
albumentations>=1.3.0
setuptools
wheel
pip
